Add docs on with clauses
