+++ examples/clauses/match.ts
@@ -17,26 +17,25 @@
  * limitations under the License.
  */
 
-import Cypher from "..";
+import Cypher from "../../dist";
 
-// MERGE (this0:MyLabel)
-// ON MATCH SET
-//     this0.count = (this0.count + 1)
-// ON CREATE SET
-//     this0.count = 1
+// MATCH (this1:`Person`)-[this0:ACTED_IN]->(this2:`Movie`)
+// WHERE (this1.name = $param0 AND this2.released = $param1)
+// RETURN this2.title, this2.released AS year
 
-const node = new Cypher.Node();
+const movieNode = new Cypher.Node();
+const personNode = new Cypher.Node();
 
-const countProp = node.property("count");
-const query = new Cypher.Merge(
-    new Cypher.Pattern(node, {
-        labels: ["MyLabel"],
-    })
-)
-    .onCreateSet([countProp, new Cypher.Literal(1)])
-    .onMatchSet([countProp, Cypher.plus(countProp, new Cypher.Literal(1))]);
+const actedInPattern = new Cypher.Pattern(movieNode, { labels: ["Movie"] })
+    .related({ type: "ACTED_IN" })
+    .to(personNode, { labels: ["Person"] });
 
-const { cypher, params } = query.build();
+const matchQuery = new Cypher.Match(actedInPattern)
+    .where(personNode, { name: new Cypher.Param("Keanu Reeves") })
+    .and(movieNode, { released: new Cypher.Param(1999) })
+    .return(movieNode.property("title"), [movieNode.property("released"), "year"]);
+
+const { cypher, params } = matchQuery.build();
 
 console.log("Cypher");
 console.log(cypher);
